Server receives and stores username
Client now sends username to server on connection, and server receives and stores it for later use. On disconnect, the username is removed. 
Usernames should be indexed in the same order as connected sockets in master fd_set array. 
Currently only uses the username in the greeting message, and in new server-side console messages.